refactor(chart): update chart component structure by replacing deprecated imports, enhancing data validation, and improving chart rendering logic for better performance
